build: bring the tree up to the current dev environment (CUDA 13.3, clang-18)
Collapses the build-system, container, environment, and toolchain
work needed to compile against the current dev container into a
single commit at the root of this branch, so every application
commit that follows builds against one consistent environment
instead of a mix of old and new build systems spread across the
original history.

Consolidates: clang-format tooling migration (clang-format-14 to
clang-format-18), CI/container Dockerfile updates, dev-container
updates, a configure.ac git-worktree-detection and unquoted-version
fix, native sm_120 SASS generation alongside the target arch, a
wx/libtorch rpath lock plus bundled CUDA-13 link fixes, a CUDA-13
port of NPP buffer-size and device-property queries, unification of
the regen scripts into regenerate_project.sh, static OpenMP
(libiomp5) linking for clang static builds, a --enable-mkl-dynamic
configure switch, devcontainer-symlink and external-repo gitignore
entries plus the standard Python-package ignore block, retirement
of scripts/containers (recipes moved to plasmon-labs-devcontainer)
and of the remote_build.sh remote-build path, removal of the old
.vscode symlink target, removal of the in-repo Claude tooling
(.claude/, CLAUDE.md files, REVIEW.md - the tools change too
quickly to maintain in the main repo), and migration of both
TemplateMatchingCore.cu and match_template.cpp to FastFFT's
PlanDescriptor-based constructor API (SetForwardFFTPlan/
SetInverseFFTPlan were removed upstream in 2026-07).

match_template.cpp mirrors the same field mapping used in
TemplateMatchingCore.cu: fwd input -> input_size, fwd output ->
fourier_size (== inv input here), inv output -> output_size, both
padded flags true (the descriptor default). FT is single-use in
this scope (only FwdFFT is called after construction), so it is
constructed directly with the plan rather than held via unique_ptr.
Without this, GPU configurations fail to build (FastFFT::
FourierTransformer's default constructor is private upstream; only
the PlanDescriptor constructor is public).

+++ src/gpu/GpuImage.cu
@@ -775,7 +775,10 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_sum:
             if ( ! is_allocated_sum_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                // CUDA 13 NPP widened the scratch-buffer-size out-parameter of the
+                // nppi*GetBufferHostSize_*_Ctx family from int* to size_t*; n_elem
+                // (a byte count passed to cudaMallocAsync) must match that width.
+                size_t n_elem;
                 nppErr(nppiSumGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->sum_buffer, n_elem, nppStream.hStream));
                 is_allocated_sum_buffer = true;
@@ -785,7 +788,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_min:
             if ( ! is_allocated_min_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMinGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->min_buffer, n_elem, nppStream.hStream));
 
@@ -796,7 +799,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_minIDX:
             if ( ! is_allocated_minIDX_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMinIndxGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->minIDX_buffer, n_elem, nppStream.hStream));
 
@@ -807,7 +810,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_max:
             if ( ! is_allocated_max_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMaxGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->max_buffer, n_elem, nppStream.hStream));
 
@@ -818,7 +821,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_maxIDX:
             if ( ! is_allocated_maxIDX_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMaxIndxGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->maxIDX_buffer, n_elem, nppStream.hStream));
 
@@ -829,7 +832,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_minmax:
             if ( ! is_allocated_minmax_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMinMaxGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->minmax_buffer, n_elem, nppStream.hStream));
 
@@ -840,7 +843,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_minmaxIDX:
             if ( ! is_allocated_minmaxIDX_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMinMaxIndxGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->minmaxIDX_buffer, n_elem, nppStream.hStream));
 
@@ -851,7 +854,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_mean:
             if ( ! is_allocated_mean_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMeanGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->mean_buffer, n_elem, nppStream.hStream));
 
@@ -861,7 +864,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_meanstddev:
             if ( ! is_allocated_meanstddev_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiMeanStdDevGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->meanstddev_buffer, n_elem, nppStream.hStream));
 
@@ -872,7 +875,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_countinrange:
             if ( ! is_allocated_countinrange_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiCountInRangeGetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->countinrange_buffer, n_elem, nppStream.hStream));
 
@@ -883,7 +886,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_l2norm:
             if ( ! is_allocated_l2norm_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiNormL2GetBufferHostSize_32f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->l2norm_buffer, n_elem, nppStream.hStream));
 
@@ -894,7 +897,7 @@ void GpuImage::BufferInit(BufferType bt, int n_elements) {
         case b_dotproduct:
             if ( ! is_allocated_dotproduct_buffer ) {
                 MyDebugAssertTrue(is_npp_loaded, "Error: NPP not loaded");
-                int n_elem;
+                size_t n_elem;
                 nppErr(nppiDotProdGetBufferHostSize_32f64f_C1R_Ctx(npp_ROI, &n_elem, nppStream));
                 cudaErr(cudaMallocAsync(&this->dotproduct_buffer, n_elem, nppStream.hStream));
 
@@ -1698,7 +1701,7 @@ _pre_GetWeightedCorrelationWithImageKernel(const __restrict__ cufftComplex* imag
                 sum3 = v;
             }
 #else
-            image_PS[real_idx] = ComplexModulusSquared(image_values[complex_idx]);
+            image_PS[real_idx]      = ComplexModulusSquared(image_values[complex_idx]);
             projection_PS[real_idx] = ComplexModulusSquared(projection_values[complex_idx]);
             if ( u > signed_CC_limit_sq ) {
                 cross_terms[real_idx] = fabsf(v);
@@ -1720,8 +1723,8 @@ _pre_GetWeightedCorrelationWithImageKernel(const __restrict__ cufftComplex* imag
         sum2 = 0.f;
         sum3 = 0.f;
 #else
-        cross_terms[real_idx] = 0.f;
-        image_PS[real_idx] = 0.f;
+        cross_terms[real_idx]   = 0.f;
+        image_PS[real_idx]      = 0.f;
         projection_PS[real_idx] = 0.f;
 #endif
     }
